+++ src/createManager.js
@@ -1,15 +1,7 @@
 /** @flow */
-import { queriesCreator, type Queries } from './manager'
-import { createRepository, type Repository } from './createRepository'
-import {
-  connect as createPool,
-  type Pool,
-  type Connection,
-} from './connectionManager'
+import { createBaseManager, type Repositories } from './manager'
 import { makeQueryBuilder } from './query'
-import createMetadataManager from './utils/metadataManager'
 import type {
-  MetadataManager,
   BaseRecord,
   InsertQuery,
   SelectQuery,
@@ -18,24 +10,7 @@ import type {
 } from './types'
 
 export type Manager = $Exact<{
-  ...Queries,
-  connect(): void,
-  getRepository<T: BaseRecord>(tableName: string): Repository<T>,
-  extendRepository(
-    tableName: string,
-    callback: (repo: Repository<>) => { ...Repository<> },
-  ): void,
-  setRepositoryFactory(
-    factory: (tableName: string, manager: Manager) => Repository<>,
-  ): void,
-  getLogger(): ?typeof console,
-  ready(): Promise<any>,
-  getPool(): Pool,
-  clear(): void,
-  getMetadataManager(): MetadataManager,
-  setMetadataManager(manager: MetadataManager): void,
-  getConnection(): Promise<Connection>,
-  getConfiguration(): { [key: string]: any },
+  ...Repositories,
   startQuery(): {
     insert(): InsertQuery,
     select<T: BaseRecord>(): SelectQuery<T>,
@@ -45,84 +20,13 @@ export type Manager = $Exact<{
   },
 }>
 
-export function createManager(
-  connectionConfig: any,
-  logger: ?typeof console = null,
-): Manager {
-  // MySQL2 Connection Pool
-  let pool: ?Pool
-  /**
-   * {
-   *   tableName: Repository
-   * }
-   */
-  let repos = {}
-  let metadataManager: MetadataManager
-  let reposirotyFactory = createRepository
-
-  function connect() {
-    if (!metadataManager) {
-      metadataManager = createMetadataManager()(this)
-    }
-    pool = createPool(connectionConfig)
-    metadataManager.ready()
-  }
-
-  function getPool() {
-    if (!pool) {
-      const msg = 'Please start connection before'
-      throw new Error(msg)
-    }
-    return pool
-  }
-
-  const queries = queriesCreator(getPool)
+export function createManager(connectionConfig: any): Manager {
+  const baseManager = createBaseManager(() => ({}))(connectionConfig)
 
   return {
-    ...queries,
-    connect,
-    async ready() {
-      return await metadataManager.ready()
-    },
-    getConfiguration() {
-      return connectionConfig
-    },
-    extendRepository(tableName, callback) {
-      repos[tableName] = callback(this.getRepository(tableName))
-    },
-    getRepository(tableName) {
-      if (!repos.hasOwnProperty(tableName)) {
-        repos[tableName] = reposirotyFactory(tableName, this)
-      }
-
-      return repos[tableName]
-    },
-    setRepositoryFactory(factory) {
-      if (typeof factory != 'function') {
-        throw new Error('Repository Factory must be a function')
-      }
-      reposirotyFactory = factory
-    },
-    getLogger() {
-      return logger
-    },
-    getPool,
-    clear() {
-      pool = null
-      repos = {}
-      metadataManager.clear()
-    },
-    getMetadataManager() {
-      return metadataManager
-    },
-    setMetadataManager(manager) {
-      metadataManager = manager
-    },
-    getConnection() {
-      return getPool().getConnection()
-    },
+    ...baseManager,
     startQuery() {
-      return makeQueryBuilder(this)
+      return makeQueryBuilder(baseManager)
     },
   }
 }
