+++ src/manager/types.js
@@ -1,0 +1,4 @@
+/** @flow */
+import type { PoolOptions } from '../connectionManager'
+
+export type ManagerConstructor<T> = (connectionConfig: PoolOptions) => T
