+++ src/utils/compose.js
@@ -1,0 +1,8 @@
+/** @flow */
+
+// $FlowIgnore
+const createComposer = (): $Compose => (...funcs) => {
+  return arg => funcs.reduceRight((composed, f) => f(composed), arg)
+}
+
+export const compose: $Compose = createComposer()
